Some of my favorite projects from DSA 1 & 2
HashTable: Implement a hash table that will go through the word search and find all of the unique words. Written in Emacs (no IDE for this class).

MergeSort: Simple mergesort written in x86 assembly. Also written in Emacs.

Trading: This is a closest pair of points problem using a divide and conquer algorithm.